Allow block listener to specify the peer listen options
1. Enable private data capability as part of peer listen
request on channel service.
2. Refactoring
